Unity 2017.1 (and more) (#246)
* Auto API update

* Update wolfulus for 2017 support

* Fixes for https://github.com/mapbox/mapbox-unity-sdk/issues/204

* Fixes: https://github.com/mapbox/mapbox-unity-sdk/issues/232

* Removing preprocessor directives for older versions of unity

* Check no longer needed

* simplified "UNITY" definition

* Fix warnings!

+++ sdkproject/Assets/Mapbox/Unity/Utilities/Conversions.cs
@@ -148,7 +148,7 @@ namespace Mapbox.Unity.Utilities
 		/// <param name="longitude"> The longitude. </param>
 		/// <param name="zoom"> Zoom level. </param>
 		/// <returns> A <see cref="T:UnityEngine.Vector2d"/> xy tile ID. </returns>
-		public static Vector2d LatitudeLongitudeToTileId(float latitude, float longitude, int zoom)
+		public static Vector2d LatitudeLongitudeToTileId(double latitude, double longitude, int zoom)
 		{
 			var x = (int)Math.Floor((longitude + 180.0) / 360.0 * Math.Pow(2.0, zoom));
 			var y = (int)Math.Floor((1.0 - Math.Log(Math.Tan(latitude * Math.PI / 180.0)
